### 2020-03-01 #2129 @labarba
👋  @karthik — The authors suggested you as editor for this submission. According to the README: "LTRpred will take any genome assembly file in fasta format as input and will generate a detailed annotation of functional and potentially mobile LTR retrotransposons." It's an R package. Can you handle it?
