### 2020-03-02 #2129 @HajkD
Dear @labarba

Thank you so much for arranging the editorial assignment.

After going through the [reviewer list](https://bit.ly/joss-reviewers), I think the following reviewers could be very suitable: rachelss and dcassol .

I very much look forward to receiving some feedback.

Many thanks!
