### 2020-03-10 #2129 @lpantano
@dcassol, would you like to review this tool about LTRpred: de novo annotation of intact retrotransposons?
