### 2020-03-10 #2129 @lpantano
Thanks @oushujun, do you know somebody who could be a reviewer for this tool? JOSS makes its reviewing process through github, and it is a very nice experience in general.
