### 2020-03-26 #2170 @whedon
[REVIEW]: LTRpred: de novo annotation of intact retrotransposons
**Submitting author:** @HajkD (<a href="http://orcid.org/0000-0002-1567-306X">Hajk-Georg Drost</a>)
**Repository:** <a href="https://github.com/HajkD/LTRpred" target ="_blank">https://github.com/HajkD/LTRpred</a>
**Version:** v1.1.0
