### 2020-04-01 HajkD/LTRpred#13 @mdozmorov
Installation instructions of dependencies should be updated
The LTRpred package requires six command-line tools, https://hajkd.github.io/LTRpred/articles/Introduction.html#installation. The installation instructions need to be updated as they refer to tools that are 2-3 versions older than the current ones. Consequently, installation instructions differ. For example:
````
wget ftp://selab.janelia.org/pub/software/hmmer3/3.1b2/hmmer-3.1b2.tar.gz
````
doesn't work. Only the latest version can be downloaded:
````
wget http://eddylab.org/software/hmmer/hmmer-3.3.tar.gz
````

Same for GenomeTools and others - installation instructions should be updated.
