### 2020-04-01 8762c0c
add biomaRt and Biostrings installation before biomartr installation in installation instructions #12 (Many thanks to Mikhail Dozmorov)
