### 2020-04-01 HajkD/LTRpred#14 @mdozmorov
Installation of dependencies fail - Docker wrapper needed
The LTRpred package requires six command-line tools. Each of them has its own dependencies and/or require `sudo` privileges. Testing with Docker containers using different OS versions failed, installation of tools breaks with various errors. Different tools break under different conditions. After hours of attempts, it seems impossible to install all the dependencies to test the package.

A Docker file containing all the dependencies and the LTRpred package is needed.
