### 2020-04-01 HajkD/LTRpred#13 @HajkD
Dear Mikhail,

Thank you for pointing this out.

Unfortunately, there are several command-line tool dependencies that don't work with
the the latest versions in each case.

In the next issue you suggested to build a Docker container. I think this is an excellent idea and I will start working on this.

Do you think a conda environment would also be useful?

Many thanks,
Hajk
