### 2020-04-01 HajkD/LTRpred#14 @HajkD
Dear Mikhail,

I am very sorry that you had difficulties to install all command line tools.

I very much like your suggestion to install all command line tools and their various dependencies within a Docker container (including LTRpred itself).

Do you think a conda environment would also be useful?

I will start working on this!
