### 2020-04-01 HajkD/LTRpred#14 @mdozmorov
Not sure it'll be possible to create a single conda recipe for so many tools. Docker seems best as it provides the relevant OS environment. I've tried https://hub.docker.com/r/rocker/rstudio to have RStudio, but again, ran in multiple issues with dependencies. A single Docker file that installs the dependencies, R, and LTRpred with all its dependencies seems the only option
