### 2020-04-02 HajkD/LTRpred#14 @HajkD
Thank you so much for this insight. In this case I will focus on the Docker file that installs everything and will come back to you shortly.

Once more, please accept my apologies for the inconvenience caused, but I hope this Docker file will make the life for many future users easier.

Many thanks!
