+++ R/ORFpred.R
@@ -98,7 +98,7 @@ ORFpred <- function(seq.file,
             is.na(file.info(orf.file)$size == 0)) {
             ReadSeqFile <- Biostrings::readDNAStringSet(seq.file)
             ORFCount.df <-
-                dplyr::data_frame(seq.id = ReadSeqFile@ranges@NAMES,
+                tibble::tibble(seq.id = ReadSeqFile@ranges@NAMES,
                                   orfs = rep(0, length(ReadSeqFile@ranges@NAMES)))
         } else {
             ORFCount.df <- read.orfs(orf.file)
@@ -109,7 +109,7 @@ ORFpred <- function(seq.file,
             is.na(file.info(orf.file)$size == 0)) {
             ReadSeqFile <- Biostrings::readDNAStringSet(seq.file)
             ORFCount.df <-
-                dplyr::data_frame(seq.id = ReadSeqFile@ranges@NAMES,
+                tibble::tibble(seq.id = ReadSeqFile@ranges@NAMES,
                                   orfs = rep(0, length(ReadSeqFile@ranges@NAMES)))
         } else {
             ORFCount.df <- read.orfs(orf.file)
