### 2020-04-05 HajkD/LTRpred#14 @HajkD
Dear Mikhail,

Please find enclosed a `Dockerfile` for `LTRpred` that will run the example `LTRpred::LTRpred(...)` function.

https://github.com/HajkD/LTRpred/blob/master/Dockerfile

This `Dockerfile` can be used to create a `ltrpred` container as follows:

```sh
# assuming the Dockerfile is in the current working directory
docker build -t ltrpred .
# run LTRpred::LTRpred(...) example via docker
docker run ltrpred
```

In the next days, I will also add an RStudio interface via `rocker` and design a way for users to interact with the `LTRpred` docker container (passing either scripts or arguments to the LTRpred run).

You were absolutely correct, during the past months new dependency conflicts emerged between different tool versions. I resolved them now in the `Dockerfile` and will also update them in the Vignette.

I hope this `Dockerfile` allows you to run `LTRpred` seamlessly and once more thank you so much for motivating this approach.

Best wishes,
Hajk
