### 2020-04-06 HajkD/LTRpred#14 @mdozmorov
Thanks, Hajk, this is a great start! The Docker image compiles, takes some time. Looking forward to the full version when one can play with LTRpred via RStudio interface.
