### 2020-04-13 #2170 @mdozmorov
The package has installation problems, the issues have been opened https://github.com/HajkD/LTRpred/issues/14, https://github.com/HajkD/LTRpred/issues/13. The review will continue when the package will be possible to test
