### 2020-04-13 #2170 @HajkD
Dear all,

with the most recent `Dockerfile` on GitHub it is now possible to run all examples shown in the manuscript. As well as playing with all `LTRpred` functions.

https://github.com/HajkD/LTRpred/blob/master/Dockerfile

This Dockerfile can be used to create a `ltrpred` Docker container as follows:

```sh
# assuming the Dockerfile is in the current working directory
docker build -t ltrpred .
# running ltrpred container
docker run --rm -ti ltrpred
# open R
:/app# R
```
Now in the R prompt:
```r
# all tools are installed and the example LTRpred::LTRpred(...) run etc can be used
LTRpred::LTRpred(genome.file = system.file("Hsapiens_ChrY.fa", package = "LTRpred"))
```

What is missing and I would like to add is the ability to copy a local genome `fasta` file to the running `ltrpred` docker image. In addition, I think it could be nice and useful for users to be able to communicate with the `ltrpred` docker image via `RStudio`. For this purpose, I am implementing an interface between the `ltrpred` docker image and `RStudio`.

The core functionality of `LTRpred`, however, can now be used.

I hope this helps to make it easier to interact with this package.

I will try to have the final interfaces (ltrpred docker/RStudio and local file copy to ltrpred docker)
running in the next few days.

Many thanks!
