### 2020-04-13 #2170 @HajkD
P.S. I will also push the final docker image to `Docker Hub` so that users can easily retrieve `LTRpred`.
