### 2020-04-16 HajkD/LTRpred#14 @mdozmorov
Following the instructions:
````
# assuming the Dockerfile is in the current working directory
docker build -t ltrpred .
# running ltrpred container
docker run --rm -ti ltrpred
# open R
:/app# R
````
Now in the R prompt:
````
# all tools are installed and the example LTRpred::LTRpred(...) run etc can be used
LTRpred::LTRpred(genome.file = system.file("Hsapiens_ChrY.fa", package = "LTRpred"))
````
I'm getting
````
LTRpred::LTRpred(genome.file = system.file("Hsapiens_ChrY.fa", package = "LTRpred"))
Error in loadNamespace(name) : there is no package called ‘LTRpred’
````
I see it is being installed in the Docker file, but something is not working. Installing within the container results in
````
ERROR: dependency ‘ggbio’ is not available for package ‘LTRpred’
* removing ‘/Users/mdozmorov/Library/R/3.6/library/LTRpred’
Error: Failed to install 'LTRpred' from GitHub:
  (converted from warning) installation of package ‘/var/folders/jq/hvjs8pl55sx0mtxlptkms_cc6y3f75/T//RtmpWeE9C6/fileed90376dfe7b/LTRpred_1.1.0.tar.gz’ had non-zero exit status
````
Some more polishing is needed.
