### 2020-04-20 HajkD/LTRpred#14 @HajkD
Dear Mikhail,

Many thanks for letting me know. I am still troubleshooting.
It seems like these packages are not yet compatible with the most recent R version.
I have to find a way around this.

The reason why it worked on my side was that parts of the previous built were stored in my Docker cache. This way, my test example worked.

Now when running:

```sh
docker build --no-cache -t ltrpred .
```

I could reproduce the issue and found out that there are issues with the following packages:

```r
Skipping 6 packages not available: GenomicRanges, GenomeInfoDb, Biostrings, IRanges, ggbio, biomaRt
```

I am very sorry for the delay, but I hope to be able to create a stable `Dockerfile` version so that I can then push it to docker hub.

Many thanks and best wishes,
Hajk
