### 2020-04-20 HajkD/LTRpred#14 @mdozmorov
Guess it can be R/Bioconductor version conflict. Will be waiting.
