### 2020-05-04 ca66578
Update general installation instructions and add docker container installation instructions (Many thanks to Mikhail Dozmorov) #13 #14
