+++ vignettes/Introduction.Rmd
@@ -12,10 +12,12 @@ vignette: >
 - [1. Introduction](#introduction)
 - [2. Getting Started](#getting-started)
     - [2.1 Installation](#installation)
-    - [2.1 Quick Start](#quick-start)
-    - [2.2 LTRpred output](#ltrpred-output)
-    - [2.3 Import LTRpred output](#import-ltrpred-output)
-    - [2.4 Output file format of LTRpred()](#output-file-format-of-ltrpred)
+        - [2.1.1 LTRpred Docker Container](#ltrpred-docker-container)
+        - [2.1.2 Install tool dependencies on Linux](#install-tool-dependencies-on-linux)
+    - [2.2 Quick Start](#quick-start)
+    - [2.3 LTRpred output](#ltrpred-output)
+    - [2.4 Import LTRpred output](#import-ltrpred-output)
+    - [2.5 Output file format of LTRpred()](#output-file-format-of-ltrpred)
 - [3. Detailed LTRpred run](#detailed-ltrpred-run)
     - [3.1 HMM Models](#hmm-models)
     - [3.2 Detailed description of adjustable LTRpred parameters](#detailed-description-of-adjustable-ltrpred-parameters)
@@ -57,166 +59,256 @@ __Figure 1__ in mind will help researchers to modify structural parameters in a 
 
 ## Installation
 
+The fastest way to run `LTRpred` is to download the `ltrpred` Docker container which
+includes all pre-installed tool dependencies. For users who cannot use a Docker environment
+the individual installation instructions for each dependency tool are listed below (for MacOS and Linux).
+
 __Note: please be aware that LTRpred is designed for Linux and Unix based systems such as macOS. Hence, although LTRpred itself can be installed on Windows machines, most of the command line tools LTRpred depends on cannot be installed on Windows machines.__
 
-### Programming Languages
+
+### LTRpred Docker Container
+
+Please make sure to create a [Docker](https://www.docker.com/get-started) account and to [install Docker](https://docs.docker.com/engine/install/) on your system.
+
+#### Download `ltrpred` container for use with R command line
+
+
+```bash
+# retrieve docker image from dockerhub
+docker pull drostlab/ltrpred
+# run ltrpred container
+docker run --rm -ti ltrpred
+# start R prompt within ltrpred container
+~:/app# R
+```
+
+Within the ltrpred container R prompt run the `ltrpred` example:
+
+```r
+LTRpred::LTRpred(genome.file = system.file("Hsapiens_ChrY.fa", package = "LTRpred"))
+```
+
+To exit R in the container run:
+
+```r
+q()
+```
+
+And to exit the `ltrpred` container run:
+
+```bash
+~:/app# exit
+```
+
+Now, users can add their own genome data as well as the Dfam database for
+further annotation to the `ltrpred` container by following these steps (in a different Terminal window):
+
+```bash
+# go to the folder path in which you want to
+# store all genome and Dfam data you want to
+# mount in the ltrpred container and then run:
+# create a new folder which will store
+# all files that will be required in the
+# ltrpred container
+mkdir ltrpred_data
+cd ltrpred_data
+# create a dfam database folder
+mkdir Dfam
+cd Dfam
+```
+
+Now users can download and format the Dfam database as follows (within the `Dfam` folder created above):
+
+__For this step users need to install [HMMER](http://hmmer.org/download.html) on their machine.__ An example installation of HMMER for Linux machines is listed below.
+
+```bash
[…]
