+++ vignettes/Introduction.Rmd
@@ -135,6 +135,7 @@ want to _de novo_ annotate with `LTRpred` in the `ltrpred_data` folder you just 
 A possible way to retrieve such a genome is (within R):
 
 ```r
+# install.packages("biomartr")
 biomartr::getGenome(db = "ensembl", 
                     organism = "Saccharomyces cerevisiae", 
                     path = "yeast_genome", 
