### 2020-05-04 23e53ab
update installation instruction
