### 2020-05-04 #2170 @HajkD
Dear all,

Please accept my apologies for being silent for a while, but I wanted to make sure that this time I can deliver a complete and interactive `ltrpred` container toolkit rather than a `Dockerfile` prototype.

As a result, I developed two types of `ltrpred` containers which are available at `Docker Hub`:

1) Running the `LTRpred` pipeline through the `R command line prompt`: https://hub.docker.com/repository/docker/drostlab/ltrpred/.
2) Running the `LTRpred` pipeline through an RStudio Server interface: https://hub.docker.com/repository/docker/drostlab/ltrpred_rstudio.

I now documented in detail how to install, interact, and run both types of docker images here:
https://hajkd.github.io/LTRpred/articles/Introduction.html#ltrpred-docker-container

A minimal example to get you started right away is the following:

```bash
# retrieve docker image from dockerhub
docker pull drostlab/ltrpred
# run ltrpred container
docker run --rm -ti drostlab/ltrpred
# start R prompt within ltrpred container
~:/app# R
```

Within the ltrpred container R prompt run the `ltrpred` example:

```r
LTRpred::LTRpred(genome.file = system.file("Hsapiens_ChrY.fa", package = "LTRpred"))
```

The RStudio variant can be found here:
https://hajkd.github.io/LTRpred/articles/Introduction.html#download-ltrpred_rstudio-container-for-use-with-rstudio-server

It took me a while to find the best way to allow users to run `ltrpred` through a `RStudio Server` interface, but I think the solution presented here should make it as easy as possible to use the `ltrpred` pipeline with `RStudio`.

I would like to express my gratitude for this brilliant suggestion and I am sure that future users will appreciate having access to a pipeline environment where all tool dependencies are preinstalled.

I am very well aware that the `ltrpred` pipeline depends on a large number of other tools, but
the new functionality and analyses that can be performed thanks to the integration of all these tools are very useful for `transposable element` research and the type of annotation and analysis output generated by `ltrpred` was urgently needed.

I now very much look forward to hearing your feedback regarding the pipeline and further ways to improve it.

Your help and efforts are very much appreciated!

Many thanks,

Hajk
