### 2020-05-04 HajkD/LTRpred#13 @HajkD
Dear Mikhail,

I now updated the installation instructions for the Linux tools using the most recent tool versions:
https://hajkd.github.io/LTRpred/articles/Introduction.html#install-tool-dependencies-on-linux

I decided to make these instructions only available for `Linux` since I am referring `macOS` and `Windows` users to the [ltrpred Docker](https://hub.docker.com/repository/docker/drostlab/ltrpred/) container.

Once more thank you very much for making me aware of this issue of the tool versions and I will make sure to keep the tools up to date in the future.

Many thanks,
Hajk
