### 2020-05-04 HajkD/LTRpred#14 @HajkD
Dear Mikhail,

I finally managed to develop two types of `ltrpred` containers:

1) Running the `LTRpred` pipeline through the `R prompt` command line: https://hub.docker.com/repository/docker/drostlab/ltrpred/.
2) Running the `LTRpred` pipeline through RStudio Server: https://hub.docker.com/repository/docker/drostlab/ltrpred_rstudio.

I now documented in detail how to install and run both types of docker images here:
https://hajkd.github.io/LTRpred/articles/Introduction.html#ltrpred-docker-container

A minimal example to get you started right away is the following:

```bash
# retrieve docker image from dockerhub
docker pull drostlab/ltrpred
# run ltrpred container
docker run --rm -ti drostlab/ltrpred
# start R prompt within ltrpred container
~:/app# R
```

Within the ltrpred container R prompt run the `ltrpred` example:

```r
LTRpred::LTRpred(genome.file = system.file("Hsapiens_ChrY.fa", package = "LTRpred"))
```

The RStudio variant can be found here:
https://hajkd.github.io/LTRpred/articles/Introduction.html#download-ltrpred_rstudio-container-for-use-with-rstudio-server

Please accept my apologies for taking a while to implement and deploy these containers, but I wanted to do it properly. I tried many different ways of how users may be able to interact with the containers (how to use local genome files or local versions of the Dfam database) and now came up with the most parsimonious way as is documented here:
https://hajkd.github.io/LTRpred/articles/Introduction.html#ltrpred-docker-container

I hope you find these new versions useful and I very much look forward to hearing your feedback.

Thank you so much for all your excellent help with this. I am sure that users will very much appreciate this new style of working with `LTRpred`.

With very best wishes,
Hajk
