### 2020-05-05 HajkD/LTRpred#13 @mdozmorov
Great work, and the software works now. Please, fix minor instruction issues, so we can close this issue:

- Add some readme to https://hub.docker.com/r/drostlab/ltrpred and its Rstudio variant
- Suggesting simplifying and streamlining Docker instructions, like:
    - Explicitly mention that the command-line version is suited for Windows, where port bridge cannot be made, so not RStudio access
    - Use -v volume mapping for the command-line version - the same type of analysis should be possible via command-line and RStudio versions. Terminal in RStudio is helpful
    - HMMER already installed within both images, so no need to install it within the containers
    - Explain why Dfam is not included (for one, it is very large)
    - Some minor typo fixes: "The rationalE for", "rertrotransposon",  "yeasT_genome"
