### 2020-05-05 HajkD/LTRpred#13 @mdozmorov
Also, recommend mentioning Docker options directly in the Git repo README. Many will want to use it
