### 2020-05-05 5b16f0f
adding more clear documentation for container usage (as suggested by Mikhail Dozmorov) #13
