+++ vignettes/Introduction.Rmd
@@ -35,7 +35,7 @@ Please make sure that all command line tools are installed properly before runni
 
 # Getting Started
 
-The rational for implementing `LTRpred` was to implement an R based pipeline combining the most
+The rationale for implementing `LTRpred` was to implement an R based pipeline combining the most
 sensitive, accurate, and conservative state-of-the art LTR retrotransposon
 detection tools and extend their inference by additional analyses and quality filtering steps to
 screen for functional and structurally intact elements. Hence, _LTRpred_
@@ -50,7 +50,7 @@ Hence, these `de novo` annotation tools are designed to screen the genome system
 these structural DNA features. __Figure 1__ shows the structural features of LTR retrotransposons
 that are used for predicting putative LTR transposons `de novo` in any genome of interest.
 
-![Sequence features of LTR retrotransposons. The structural element (LTR length, LTR similarity, TSD, max. size of full LTR rertrotransposon, PBS length, tRNA binding, protein domain search, etc.) can be modified and controlled separately by corresponding arguments implemented in the `LTRpred()` function. In addition to controlling the structural features of candidate LTR retrotransposons, a open reading frame prediction, Dfam annotation, copy number clustering, and LTR copy number estimation, and methylation context quantification is performed subsequently after reporting putative LTR retrotransposons.](LTRfeatures.png)
+![Sequence features of LTR retrotransposons. The structural element (LTR length, LTR similarity, TSD, max. size of full LTR retrotransposon, PBS length, tRNA binding, protein domain search, etc.) can be modified and controlled separately by corresponding arguments implemented in the `LTRpred()` function. In addition to controlling the structural features of candidate LTR retrotransposons, a open reading frame prediction, Dfam annotation, copy number clustering, and LTR copy number estimation, and methylation context quantification is performed subsequently after reporting putative LTR retrotransposons.](LTRfeatures.png)
 
 Based on the optimized output of these tools, the `LTRpred` package aims to provide researchers with maximum flexibility of
 adjustable parameters to detect any type of functional LTR retrotransposon. `LTRpred` package allows users to modify a vast range of
@@ -65,9 +65,11 @@ the individual installation instructions for each dependency tool are listed bel
 
 ### LTRpred Docker Container
 
+Please be aware that the `drostlab/ltrpred` container (command line version) is suitable for `Linux`, `macOS`, and `Windows` users. Whereas the `RStudio` container version `drostlab/ltrpred_rstudio` is not suitable for `Windows` users, because a port bridge cannot be established.  
+
 Please make sure to create a [Docker](https://www.docker.com/get-started) account and to [install Docker](https://docs.docker.com/engine/install/) on your system.
 
-#### Download `ltrpred` container for use with R command line
+#### Download `drostlab/ltrpred` container for use with R command line
 
 
 ```bash
@@ -114,25 +116,25 @@ mkdir Dfam
 cd Dfam
 ```
 
-Now users can download and format the Dfam database as follows (within the `Dfam` folder created above):
[…]
