### 2020-05-05 HajkD/LTRpred#13 @HajkD
Dear Mikhail,

I am very happy to hear that both containers are running on your machine.

Thank you for spotting the typos and for your suggestions to improve the documentation.

As suggested, I now added README files to https://hub.docker.com/r/drostlab/ltrpred and https://hub.docker.com/r/drostlab/ltrpred_rstudio .

In addition, I addressed all your points in the Vignette and pushed the updates to the repository:

- Explicitly mention that the command-line version is suited for Windows, where port bridge cannot be made, so not RStudio access
- Use -v volume mapping for the command-line version - the same type of analysis should be possible via command-line and RStudio versions. Terminal in RStudio is helpful
- HMMER already installed within both images, so no need to install it within the containers
- Explain why Dfam is not included (for one, it is very large)
- Some minor typo fixes: "The rationalE for", "rertrotransposon", "yeasT_genome"

Thank you very much for your help with this!

Best wishes,
Hajk
