### 2020-05-08 #2170 @mdozmorov
The software LTRpred software for annotation of retrotransposons within genomic sequences is fully functional. The author should be applauded for the amount of effort to make the software accessible to users. Besides the original purpose, the GitHub repository contains good examples of complex Docker files that automate the installation of dependencies and enable the use of the software both in the command line and via the RStudio interface. The documentation is up-to-date. The manuscript accurately describes the functionality of LTRpred. @whedon accept
