+++ R/LTRpred.R
@@ -27,12 +27,12 @@
 #'  by \code{suffixerator} for the use of \code{LTRharvest}. This often can be used in case the suffix file was previously 
 #'  generated, e.g. during a previous call of this function. In this case the suffix array index
 #'  file does not need to be re-computed for new analyses. This is particularly useful when 
-#'  running \code{LTRpred} with different parameter settings.
+#'  running \code{LTRpred} with different parameter settings. A possible index file name could be \code{BASENAME_index.fsa}.
 #' @param index.file.digest specify the name of the enhanced suffix array index file that is computed
 #'  by \code{suffixerator} for the use of \code{LTRdigest}. This often can be used in case the suffix file was previously 
 #'  generated, e.g. during a previous call of this function. In this case the suffix array index
 #'  file does not need to be re-computed for new analyses. This is particularly useful when 
-#'  running \code{LTRpred} with different parameter settings.
+#'  running \code{LTRpred} with different parameter settings. A possible index file name could be \code{BASENAME_index_ltrdigest.fsa}.
 #' @param LTRdigest.gff path to the LTRdigest generated GFF file, in case LTRdigest files were pre-computed previously.
 #' @param LTRpred.folder name/path of/to an existing LTRpred folder.
 #' @param tabout.file path to the LTRdigest generated tabout file file, in case LTRdigest files were pre-computed previously.
@@ -49,7 +49,7 @@
 #' @param cluster shall predicted transposons be clustered with \code{\link{CLUSTpred}}.
 #' @param clust.sim cluster reject if sequence similarity is lower than this threshold when performing clustering with \code{\link{CLUSTpred}}.
 #' @param clust.file file path to pre-computed clustering file generated by \code{\link{CLUSTpred}}.
-#' @param copy.number.est shall copy number estimation (including solo LTR prediction) be performed? Default is \code{copy.number.est = TRUE}.
+#' @param copy.number.est shall copy number estimation (including solo LTR prediction) be performed? Default is \code{copy.number.est = FALSE}.
 #' @param fix.chr.name shall chromosome names be fixed?
 #' @param cn.eval evalue for copy number estimation (BLAST hit threshold).
 #' @param range define the genomic interval in which predicted LTR transposons shall be reported
@@ -184,7 +184,7 @@
 #' \item *_BetweenLTRSeqs.fsa : DNA sequences predicted LTR retrotransposons, in particular of the region between the LTRs in fasta format. 
 #' \item *_Details.tsv : A spread sheet containing detailed information about the predicted LTRs.
 #' \item *_FullLTRRetrotransposonSeqs.fsa : DNA sequences of the entire predicted LTR retrotransposon.
-#' \item *_index.fsa : The suffixarray index file used to predict putative LTR retrotransposons.
+#' \item *_index.fsa.* : The suffixarray index file used to predict putative LTR retrotransposons.
 #' \item *_Prediction.gff : A spread sheet containing detailed additional information about the predicted LTRs (partially redundant with the *_Details.tsv file).
 #' \item *_index_ltrdigest.fsa : The suffixarray index file used to predict putative LTR retrotransposonswith \code{LTRdigest}.
 #' \item *_LTRdigestPrediction.gff : A spread sheet containing detailed information about the predicted LTRs.
@@ -207,7 +207,7 @@
 #' \item *_ORF_prediction_aa.fsa : Stores the predicted open reading frames within the predicted LTR transposons as protein sequence.
 #' \item *_LTRpred.gff : Stores the \code{LTRpred} predicted LTR transposons in GFF format.
 #' \item *_LTRpred.bed : Stores the \code{LTRpred} predicted LTR transposons in BED format. 
[…]
