+++ man/LTRpred.Rd
@@ -108,13 +108,13 @@ LTRpred(
 by \code{suffixerator} for the use of \code{LTRharvest}. This often can be used in case the suffix file was previously 
 generated, e.g. during a previous call of this function. In this case the suffix array index
 file does not need to be re-computed for new analyses. This is particularly useful when 
-running \code{LTRpred} with different parameter settings.}
+running \code{LTRpred} with different parameter settings. A possible index file name could be \code{BASENAME_index.fsa}.}
 
 \item{index.file.digest}{specify the name of the enhanced suffix array index file that is computed
 by \code{suffixerator} for the use of \code{LTRdigest}. This often can be used in case the suffix file was previously 
 generated, e.g. during a previous call of this function. In this case the suffix array index
 file does not need to be re-computed for new analyses. This is particularly useful when 
-running \code{LTRpred} with different parameter settings.}
+running \code{LTRpred} with different parameter settings. A possible index file name could be \code{BASENAME_index_ltrdigest.fsa}.}
 
 \item{LTRdigest.gff}{path to the LTRdigest generated GFF file, in case LTRdigest files were pre-computed previously.}
 
@@ -144,7 +144,7 @@ with \code{annotate = "Dfam"}.}
 
 \item{clust.file}{file path to pre-computed clustering file generated by \code{\link{CLUSTpred}}.}
 
-\item{copy.number.est}{shall copy number estimation (including solo LTR prediction) be performed? Default is \code{copy.number.est = TRUE}.}
+\item{copy.number.est}{shall copy number estimation (including solo LTR prediction) be performed? Default is \code{copy.number.est = FALSE}.}
 
 \item{fix.chr.name}{shall chromosome names be fixed?}
 
@@ -309,7 +309,7 @@ The \code{LTRpred} function generates the following output files:
 \item *_BetweenLTRSeqs.fsa : DNA sequences predicted LTR retrotransposons, in particular of the region between the LTRs in fasta format. 
 \item *_Details.tsv : A spread sheet containing detailed information about the predicted LTRs.
 \item *_FullLTRRetrotransposonSeqs.fsa : DNA sequences of the entire predicted LTR retrotransposon.
-\item *_index.fsa : The suffixarray index file used to predict putative LTR retrotransposons.
+\item *_index.fsa.* : The suffixarray index file used to predict putative LTR retrotransposons.
 \item *_Prediction.gff : A spread sheet containing detailed additional information about the predicted LTRs (partially redundant with the *_Details.tsv file).
 \item *_index_ltrdigest.fsa : The suffixarray index file used to predict putative LTR retrotransposonswith \code{LTRdigest}.
 \item *_LTRdigestPrediction.gff : A spread sheet containing detailed information about the predicted LTRs.
@@ -332,7 +332,7 @@ the HMM matches to the LTR retrotransposon candidates.
 \item *_ORF_prediction_aa.fsa : Stores the predicted open reading frames within the predicted LTR transposons as protein sequence.
 \item *_LTRpred.gff : Stores the \code{LTRpred} predicted LTR transposons in GFF format.
 \item *_LTRpred.bed : Stores the \code{LTRpred} predicted LTR transposons in BED format. 
-\item *_LTRpred_DataSheet.csv : Stores the output table as data sheet.
+\item *_LTRpred_DataSheet.tsv : Stores the output table as data sheet.
 }
 The ' * ' is an place holder for the name of the input genome file.
 
