### 2020-06-05 HajkD/LTRpred#17 @dcassol
Small details on the vignette
Dear @HajkD,
Congratulations on the software! I only have three minor points to comment:

1.
https://github.com/HajkD/LTRpred/blob/26bda714253be86c26ac3b178a5da14003b6c4e0/vignettes/Introduction.Rmd#L154
Update to:` docker run --rm -p 8787:8787 -v /put/here/your/path/to/ltrpred_data:/app/ltrpred_data -ti drostlab/ltrpred`

2.
https://github.com/HajkD/LTRpred/blob/26bda714253be86c26ac3b178a5da14003b6c4e0/vignettes/Introduction.Rmd#L183
```
[1] "Successful job 1 ."
Warning message:
`data_frame()` is deprecated as of tibble 1.1.0.
Please use `tibble()` instead.
```
Just a suggestion to update to `tibble()`.

3.
https://github.com/HajkD/LTRpred/blob/26bda714253be86c26ac3b178a5da14003b6c4e0/vignettes/Introduction.Rmd#L262
You want to add the link to the section here.

Also, I think you could provide a few details/instructions on how to copy the results from the container to the host.
