### 2020-06-05 #2170 @dcassol
I made small comments here: HajkD/LTRpred#17
@lpantano How can I make comments on the article? Find a typo.
