### 2020-06-05 #2170 @lpantano
you can open an issue there as well since the source of the paper is in their repo: https://github.com/HajkD/LTRpred/blob/master/paper.md
