### 2020-06-06 #2170 @dcassol
LTRpred pipeline focuses on the annotation of retrotransposons and retrieving structurally intact elements within any sequence of genomes. The software fulfills all the required checkboxes for the review. In particular, I am impressed with the availability of the Docker container for the package which makes the installation and running process much more easer, special for users with less programming skills. The author also provides the possibility of deploying the software both in the command line and via the RStudio interface. Also, the paper is well written and the code is clear, easy to use, and properly documented.
