### 2020-06-06 bfee290
Fixing docker call from * -ti ltrpred to * -ti drostlab/ltrpred (Many thanks to Dani Cassol) #17
