### 2020-06-08 #2170 @lpantano
Hi  @HajkD, let me know when the issues are addressed and closed so I can move forward with the acceptance :)
