### 2020-06-09 fa29cdb
fixing typo in manuscript #18
