+++ paper.md
@@ -30,7 +30,7 @@ Transposable elements (TEs) play a crucial role in altering the genomic landscap
 
 Nowadays, the _de novo_ detection of transposable elements is performed by annotation tools specifically designed to capture any type of repeated sequence, TE family, or remnant DNA loci that can be associated with known transposable elements within a genome assembly. The main goal of such efforts is to retrieve a maximum number of loci that can be associated with known TEs. If successful, such annotation can then be used to mask host genomes from TE remnants to simplify genomics studies focusing on host genes. Therefore, there is no automatically performed distinction between complete and potentially active TE and their mutated copies.
 
-Here, we introduce the ``LTRpred`` pipeline which allows to _de novo_ annotate functional and thuse potentially mobile retrotransposons in any given genome assembly. Different from other annotation tools, ``LTRpred`` focuses on retrieving structurally intact elements within sequences of genomes rather than characterizing all traces of historic TE activity. 
+Here, we introduce the ``LTRpred`` pipeline which allows to _de novo_ annotate functional and thus potentially mobile retrotransposons in any given genome assembly. Different from other annotation tools, ``LTRpred`` focuses on retrieving structurally intact elements within sequences of genomes rather than characterizing all traces of historic TE activity. 
 
 Such functional annotation is most useful when trying to spot retrotransposons responsible for recent reshuffling of genetic material in the tree of life. Detecting and further characterisation of those active retrotransposons yields the potential to harness them as mutagenesis agents by inducing transposition bursts in a controlled fashion to stimulate genomic reshaping processes towards novel traits. 
 
