### 2020-06-09 638c505
fix another typo in manuscript
