### 2020-06-09 c95637e
add documentation about how to retrieve ltrpred data from docker container #17
