### 2020-06-09 HajkD/LTRpred#17 @HajkD
Dear Dani,

thank you so much for these brilliant suggestions!

- Point 1.) I fixed the path name: https://github.com/HajkD/LTRpred/commit/bfee290e61b9041839ff67f2d891bb69c46f34ef .
- Point 2.) I fixed the depreciated `tibble::data_frame()` call here: https://github.com/HajkD/LTRpred/commit/f9054205c3cd8ff7367958970af82ac8f920fe51
The changes will go live with the new container version of `LTRpred`.
- Point 3.) As suggested, I added the link here: https://github.com/HajkD/LTRpred/commit/fe54f08a81f52a4e41e54107874c96e2fc647f0d
- Point 4.) I added a section to guide users for who wish to retrieve the `LTRpred` results from their container back to their local machine: https://github.com/HajkD/LTRpred/commit/c95637ecb23fd56d4f1330b8c04201fcdf7beea6.

I hope that everything is according to what you had in mind?

Your help with improving `LTRpred` is truly appreciated!

Many thanks,
Hajk
