### 2020-06-09 #2170 @HajkD
Dear Lorena,

Thank you so much for handling the editorial process of the `LTRpred` manuscript despite the ongoing Covid19 pandemic. I very much appreciate your efforts to keep the review process as smoothly as possible.

I now addressed all remaining suggestions raised by Dani and would wait for her approval and closure of the last issue.

Dear  Mikhail and Dani,

I am truly grateful for all your excellent suggestions and for catching all shortcomings of the `LTRpred` functionality.

I am convinced that your efforts have vastly increased the accessibility of `LTRpred` to a much wider life science audience and it taught me to always go the extra mile from now on and build a container for all my tools.

I hope that the review process didn't disrupt your daily adjustments to the Covid19 pandemic too much and I will always be grateful to all of you for your continued service to the academic cycle.

With the warmest wishes,
Hajk
