### 2020-06-10 54bc905
adding udocker option to run LTRpred docker container in a conda environment (Many thanks to Ilja Bezrukov) #16
