### 2020-06-11 #2170 @HajkD
Brilliant! Thank you very much!

I can confirm that the current version of `LTRpred` is `v1.1.0`.
