### 2020-06-12 #2170 @HajkD
Dear @arfon

Thank you for generating the article proof. I can confirm that the text and formatting are correct and as they are supposed to be. Just a very quick question: Would it be possible to rotate __Figure 1__? Is there a way I can control the figure rotation in the Markdown file? I was trying a few LateX tricks, but they didn't work. This isn't absolutely necessary, but would make the manuscript as flawless as possible.

Your help with this is truly appreciated!

Many thanks,
Hajk
