### 2020-06-12 #2170 @arfon
Hrmm, I'm not sure sorry @HajkD. I don't think it's possible to rotate with Markdown/Pandoc. Perhaps it has something to do with the figure being a PDF? Could you try converting it to a PNG and see if that preserves the correct rotation?

BTW, you can experiment with paper compilation either here (which creates some notification noise for all parties) or using our preview service at http://whedon.theoj.org
