### 2020-06-13 #2170 @HajkD
Dear @arfon

The trick worked like a charm!

Thank you so much for your help with this!

I added the PNG figure change to the Markdown file. The manuscript is perfect now.

Many thanks,
Hajk
