### 2020-06-15 #2170 @lpantano
@HajkD, can you make a Zenodo archive and send me the link. The title and the authors should match the paper information. I think you can sync Zenodo with the releases and it will create one archive link for each release. Thanks!
