### 2020-06-16 #2170 @HajkD
Dear @lpantano

I just created a LTRpred release for the GitHub repo and created a Zenodo archive.

The corresponding link is as follows:

https://doi.org/10.5281/zenodo.3895754

Thank you very much for arranging the next steps!

With very best wishes,
Hajk
