### 2020-06-16 #2170 @lpantano
thank you. But I think we need an archive without tar the content. It should look like this: https://zenodo.org/record/3755304
