### 2020-06-17 #2170 @HajkD
Dear Lorena,

Thank you very much for making me aware of this.

In my last JOSS paper I also submitted a `*.tar.gz` file, since this is the source format for R packages:

https://zenodo.org/record/1286221#.XunyeZ4zahc

Would it be possible to check wether there is a way to use `*.tar.gz` files in JOSS, because the Zenodo policy states:

`Note: File addition, removal or modification are not allowed after an upload has been published. This is because a Digital Object Identifier (DOI) is registered with  DataCite for each upload.`

Thank you very much for your help with this and I apologise if I have caused any additional work.

With very best wishes,
Hajk
