### 2020-06-18 #2170 @lpantano
Only thing to double-check whether the archive is valid being an tar file by the board of editors.
