### 2020-06-18 #2170 @arfon
@HajkD - this PR should fix the issue with your paper: https://github.com/HajkD/LTRpred/pull/19
