### 2020-06-18 #2170 @arfon
@HajkD - looks like most of your references need a `doi` field added. Can you check the suggestions from Whedon above and if they are correct add them?

Also, just as a heads-up, these should be formatted without the `https://doi.org/` bit, i.e. the BibTeX field should look like:

```
doi=10.1038/s41467-017-01374-x
```
