+++ paper.bib
@@ -26,6 +26,7 @@
   pages    = "D81--9",
   month    =  jan,
   year     =  2016,
+  doi      = {10.1093/nar/gkv1272},
   keywords = "LTRpred Manuscript",
   language = "en"
 }
@@ -62,11 +63,12 @@
   month    =  jul,
   year     =  2011,
   keywords = "LTRpred Manuscript",
+  doi      = {10.1093/nar/gkr367},
   language = "en"
 }
 
 @ARTICLE{Rognes2016-sk,
-  title    = "{VSEARCH}: a versatile open source tool for metagenomics",
+  title    = "VSEARCH: a versatile open source tool for metagenomics",
   author   = "Rognes, Torbj{\o}rn and Flouri, Tom{\'a}{\v s} and Nichols, Ben
               and Quince, Christopher and Mah{\'e}, Fr{\'e}d{\'e}ric",
   abstract = "BACKGROUND: VSEARCH is an open source and free of charge
@@ -116,6 +118,7 @@
   pages    = "e2584",
   month    =  oct,
   year     =  2016,
+  doi      = {10.7717/peerj.2584},
   keywords = "Alignment; Chimera detection; Clustering; Dereplication; Masking;
               Metagenomics; Parallellization; Searching; Sequences;
               Shuffling;LTRpred Manuscript",
@@ -145,6 +148,7 @@
   pages    = "2460--2461",
   month    =  oct,
   year     =  2010,
+  doi      = {10.1093/bioinformatics/btq461},
   keywords = "LTRpred Manuscript",
   language = "en"
 }
@@ -183,6 +187,7 @@
   pages    = "645--656",
   month    =  may,
   year     =  2013,
+  doi      = {10.1109/tcbb.2013.68},
   keywords = "LTRpred Manuscript",
   language = "en"
 }
@@ -221,6 +226,7 @@
   pages    = "7002--7013",
   month    =  nov,
   year     =  2009,
+  doi      = {10.1093/nar/gkp759},
   keywords = "LTRpred Manuscript",
   language = "en"
 }
@@ -267,6 +273,7 @@
   pages    = "18",
   month    =  jan,
   year     =  2008,
+  doi      = {10.1186/1471-2105-9-18},
   keywords = "LTRpred Manuscript",
   language = "en"
 }
@@ -302,6 +309,7 @@
   publisher = "Springer",
   month     =  jun,
   year      =  2016,
+  doi       = {10.1007/978-3-319-24277-4},
   keywords  = "LTRpred Manuscript",
   language  = "en"
 }
@@ -330,6 +338,7 @@
   pages    = "e1003118",
   month    =  aug,
   year     =  2013,
+  doi      = {10.1371/journal.pcbi.1003118},
   keywords = "LTRpred Manuscript",
   language = "en"
 }
@@ -345,6 +354,7 @@
   number   =  43,
   pages    = "1686",
   year     =  2019,
+  doi      = {10.21105/joss.01686},
   keywords = "LTRpred Manuscript"
 }
 
@@ -371,6 +381,7 @@
   pages    = "71--86",
   month    =  feb,
   year     =  2017,
+  doi      = {10.1038/nrg.2016.139},
   keywords = "LTRpred Manuscript",
   language = "en"
 }
@@ -427,6 +438,7 @@
   pages    = "e1008370",
   month    =  sep,
   year     =  2019,
+  doi      = {10.1371/journal.pgen.1008370},
   keywords = "DrostLab;LTRpred Manuscript",
   language = "en"
 }
@@ -455,6 +467,7 @@
   pages    = "3382--3392",
   month    =  dec,
   year     =  2019,
+  doi      = {10.1093/gbe/evz255},
[…]
