+++ paper.bib
@@ -58,7 +58,7 @@
               them.",
   journal  = "Nucleic Acids Res.",
   volume   =  39,
-  number   = "Web Server issue",
+  number   = "",
   pages    = "W29--37",
   month    =  jul,
   year     =  2011,
