### 2020-06-18 #2170 @HajkD
Dear @arfon

I now added `doi` fields to all references in the `bib` file and tested all the doi links. They are all correct and after checking with http://whedon.theoj.org it seems like the manuscript is in excellent shape.

Thank you very much for catching these issues!
