### 2020-06-19 #2170 @HajkD
Dear all,

Thank you very much for all your efforts. As always it has been a pleasure submitting software to JOSS. Please keep up the excellent work and I very much look forward to collaborating with you on future projects.

With very best wishes,
Hajk
